Extract templates function to own file with docs
